backend/requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0

